+++ utils/types/src/index.ts
@@ -1,4 +1,4 @@
-export * from './remote';
-export * from './remote-utils';
+export * from './remote-events';
+export * from './remote-connection';
 export * from './chart';
 export * from './poll';
